Hot fix in Company Web service and rename Multi computer remove improved

+++ persistence/src/test/java/com/excilys/formation/cdb/persistence/TestDAO.java
@@ -136,7 +136,7 @@ public class TestDAO {
 
 	@Test(expected = DAOException.class)
 	public void invalidRemove() {
-		computerDAO.remove(null);
+		computerDAO.remove((Long) null);
 	}
 
 	@Test(expected = DAOException.class)
